src/test/javascript/spec/app/entities/wo-customs-brokerage/wo-customs-brokerage.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { WoCustomsBrokerageService } from 'app/entities/wo-customs-brokerage/wo-customs-brokerage.service';
import { IWoCustomsBrokerage, WoCustomsBrokerage } from 'app/shared/model/wo-customs-brokerage.model';

describe('Service Tests', () => {
    describe('WoCustomsBrokerage Service', () => {
        let injector: TestBed;
        let service: WoCustomsBrokerageService;
        let httpMock: HttpTestingController;
        let elemDefault: IWoCustomsBrokerage;
        let expectedResult;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            expectedResult = {};
            injector = getTestBed();
            service = injector.get(WoCustomsBrokerageService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new WoCustomsBrokerage(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
        });

        describe('Service methods', () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: elemDefault });
            });

            it('should create a WoCustomsBrokerage', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new WoCustomsBrokerage(null))
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: expected });
            });

            it('should update a WoCustomsBrokerage', async () => {
                const returnedFromService = Object.assign(
                    {
                        monthlyImportShipment: 'BBBBBB',
                        monthlyExportShipment: 'BBBBBB',
                        shipmentValue: 'BBBBBB',
                        productInfo: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => (expectedResult = resp));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(returnedFromService);
                expect(expectedResult).toMatchObject({ body: expected });
            });

            it('should return a list of WoCustomsBrokerage', async () => {
                const returnedFromService = Object.assign(
                    {
                        monthlyImportShipment: 'BBBBBB',
                        monthlyExportShipment: 'BBBBBB',
                        shipmentValue: 'BBBBBB',
                        productInfo: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => (expectedResult = body));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush([returnedFromService]);
                httpMock.verify();
                expect(expectedResult).toContainEqual(expected);
            });

            it('should delete a WoCustomsBrokerage', async () => {
                const rxPromise = service.delete(123).subscribe(resp => (expectedResult = resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
                expect(expectedResult);
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
